demo.py
"""
ResGRUU-Net 快速演示脚本
展示如何使用实现的代码进行训练和预测
"""

import os
import torch
import numpy as np
from PIL import Image

def quick_demo():
    """快速演示"""
    print("=" * 60)
    print("ResGRUU-Net 快速演示")
    print("=" * 60)
    
    # 1. 检查环境
    print("\n1. 环境检查")
    print("-" * 30)
    print(f"PyTorch版本: {torch.__version__}")
    print(f"CUDA可用: {torch.cuda.is_available()}")
    if torch.cuda.is_available():
        print(f"GPU设备: {torch.cuda.get_device_name(0)}")
    
    # 2. 创建模型
    print("\n2. 创建模型")
    print("-" * 30)
    from model import ResGRUUNet, count_parameters
    model = ResGRUUNet(in_channels=3, out_channels=1, init_features=64)
    print(f"模型创建成功")
    print(f"参数量: {count_parameters(model):,}")
    
    # 3. 模型结构概览
    print("\n3. 模型结构概览")
    print("-" * 30)
    print("ResGRUU-Net 包含以下主要组件:")
    print("- 4个下采样块 (DownBlock)")
    print("- 1个瓶颈层 (Bottleneck)")
    print("- 4个上采样块 (UpBlock)")
    print("- 每个卷积块包含2-3个ResGRU块")
    print("- ResGRU块包含2个GRU模块 + 残差连接")
    
    # 4. 前向传播演示
    print("\n4. 前向传播演示")
    print("-" * 30)
    # 创建随机输入
    batch_size = 1
    input_tensor = torch.randn(batch_size, 3, 48, 48)
    print(f"输入形状: {input_tensor.shape}")
    
    # 前向传播
    output_tensor = model(input_tensor)
    print(f"输出形状: {output_tensor.shape}")
    print(f"输出范围: [{output_tensor.min():.3f}, {output_tensor.max():.3f}]")
    
    # 5. 训练流程说明
    print("\n5. 训练流程说明")
    print("-" * 30)
    print("完整训练步骤:")
    print("1. 准备DRIVE数据集 (40幅视网膜图像)")
    print("2. 数据预处理: 灰度变换、标准化、CLAHE等")
    print("3. 数据增强: 随机切片 + Mosaic切片")
    print("4. 使用SGD优化器训练150个epoch")
    print("5. 监控指标: 准确率、AUC、敏感度、特异性")
    print("6. 自动保存最佳模型和检查点")
    
    # 6. 使用示例
    print("\n6. 训练命令示例")
    print("-" * 30)
    print("基本训练:")
    print("  python main.py --data_dir ../DRIVE --epochs 150 --batch_size 32")
    print("\n恢复训练:")
    print("  python main.py --data_dir ../DRIVE --resume ./results/checkpoints/epoch_50.pth")
    print("\n仅测试模式:")
    print("  python main.py --data_dir ../DRIVE --test_only")
    
    # 7. 预期性能
    print("\n7. 预期性能 (DRIVE数据集)")
    print("-" * 30)
    print("根据论文，ResGRUU-Net 预期性能:")
    print("- 准确率 (AC): 95.59%")
    print("- AUC: 0.9784")
    print("- 敏感度 (SE): 0.7664")
    print("- 特异性 (SP): 0.9927")
    print("- 训练时间相比R2U-Net有明显缩短")
    
    print("\n" + "=" * 60)
    print("演示完成！请确保DRIVE数据集路径正确后开始训练。")

if __name__ == "__main__":
    quick_demo()